+++ firmware/app/driver/bat.cpp
